Remove 'double' conversion

+++ 0x00-math_complex/7-division.c
@@ -11,10 +11,10 @@
 
 void division(complex c1, complex c2, complex *c3)
 {
-	double a = (double)(c1.re);
-	double b = (double)(c1.im);
-	double p = (double)(c2.re);
-	double q = (double)(c2.im);
+	double a = c1.re;
+	double b = c1.im;
+	double p = c2.re;
+	double q = c2.im;
 
 	c3->re = ((a * p) + (b * q)) / (pow(p, 2) + pow(q, 2));
 	c3->im = ((b * p) + (a * q)) / (pow(p, 2) + pow(q, 2));
